restructure meal/ingredient scripts to reduce coupling
