+++ frontend/www/meal.html
@@ -69,10 +69,10 @@
           <div class="mb-2 p-2 border border-secondary rounded">
             <div class="text-center fw-bold">Nutrition Totals</div>
             <div class="d-flex align-items-center justify-content-between">
-              <div id="meal-ingredient-calories-total" class="me-1"></div>
-              <div id="meal-ingredient-protein-total" class="me-1"></div>
-              <div id="meal-ingredient-fat-total" class="me-1"></div>
-              <div id="meal-ingredient-carbohydrates-total"></div>
+              <div class="meal-ingredient-calories-total me-1"></div>
+              <div class="meal-ingredient-protein-total me-1"></div>
+              <div class="meal-ingredient-fat-total me-1"></div>
+              <div class="meal-ingredient-carbohydrates-total"></div>
             </div>
           </div>
           <div id="meal-ingredient-input-container">
