+++ frontend/www/meals.html
@@ -67,25 +67,11 @@
               </tr>
             </table>
           </div>
-          <div class="flex-grow-1">
-            <table class="mb-1 text-start right-table">
-              <tr>
-                <td class="pe-1">Calories:</td>
-                <td></td>
-              </tr>
-              <tr>
-                <td class="pe-1">Protein:</td>
-                <td></td>
-              </tr>
-              <tr>
-                <td class="pe-1">Fat:</td>
-                <td></td>
-              </tr>
-              <tr>
-                <td class="pe-1">Carbohydrates:</td>
-                <td></td>
-              </tr>
-            </table>
+          <div class="flex-grow-1 text-start">
+            <div class="meal-ingredient-calories-total"></div>
+            <div class="meal-ingredient-protein-total"></div>
+            <div class="meal-ingredient-fat-total"></div>
+            <div class="meal-ingredient-carbohydrates-total"></div>
           </div>
         </div>
       </a>
